Change the way we quote for fish
Specifically, inert characters can go inside or outside of quotes. Fish is a
little bit different to the other shells in that we need to move in and out of
quoted sections. When we're printing inert characters, we can do so regardless
of whether we're in a quoted section or not.