Change quoting strategy for `sh`
This reflects a mistake in the previous strategy where I had validated the
approach with roundtrips through `sh` using the `echo` builtin. The `echo`
builtin does escape processing _in addition_ to whatever the shell does.
Roundtrips via `echo` were therefore almost entirely invalid.

The combination of this with next to no documentation about how to escape
strings in /bin/sh led to a useless implementation.

Dash, I have discovered, is a sort of spiritual successor to the /bin/sh of
lore, and does have some documentation on how to escape strings. The
implementation herein is based on that documentation.

It also uses `printf %s` for roundtrips, which does _not_ perform additional
escape processing.