Slim down package to be published
Previously, things like the `fish` subdirectory were being included, which
contained only binaries of `fish` against which I was running tests locally.
These are not useful artifacts to publish to crates.io. Indeed, nor are the
tests and benchmarks; these are published as part of the source repository of
course, however.